requirements.txt
amulet-nbt==2.1.5
dfpyre==0.8.17
mcitemlib==0.4.4
mutf8==1.0.6
numpy==1.26.4
result==0.17.0
websocket-client==1.8.0
